examples/01-llm-call-open-ai/pyproject.toml
#
# Copyright 2026 The Dapr Authors
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#     http://www.apache.org/licenses/LICENSE-2.0
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

[project]
name = "01-llm-call-open-ai"
version = "0.1.0"
requires-python = ">=3.11, <3.14"
dependencies = [
    "dapr-agents",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
dapr-agents = { workspace = true }
